feat: add radio

+++ Streamlit/demo.py
@@ -64,7 +64,7 @@ def main():
     if img_file is not None:
         image = Image.open(img_file)
 
-        st.image(image, use_column_width=True)
+        st.image(image)
 
         if "points" not in st.session_state:
             st.session_state["points"] = []
@@ -88,7 +88,7 @@ def main():
                     coords = get_ellipse_coords(point)
                     draw.ellipse(coords, fill="red")
 
-                value = streamlit_image_coordinates(img, width=350, key="pil")
+                value = streamlit_image_coordinates(img, key="pil")
 
                 if value is not None:
                     point = value["x"], value["y"]
@@ -103,34 +103,37 @@ def main():
 
     st.subheader("Annotation Tool")
 
-    zip_file = st.file_uploader("사진 데이터 파일 .zip 형식으로 업로드(윈도우용)", type=["zip"])
-    
-    if zip_file is not None and is_folder_empty("images_win"):
-        with zipfile.ZipFile(zip_file, "r") as zip_ref:
-            zip_ref.extractall("images_win")
-    
+    option = st.radio('Select one:', ["Window", "MAC"])
 
-    zip_file = st.file_uploader("사진 데이터 파일 .zip 형식으로 업로드(Mac용)", type=["zip"])
-    
-    if zip_file is not None and is_folder_empty("images_mac"):
-        with zipfile.ZipFile(zip_file, "r") as zip_ref:
-            zip_ref.extractall("images_mac")
+    if option == "Window":
+        zip_file = st.file_uploader("사진 데이터 파일 .zip 형식으로 업로드(윈도우용)", type=["zip"])
         
-        macosx_folder = "images_mac/__MACOSX"
-        if os.path.exists(macosx_folder) and os.path.isdir(macosx_folder):
-            shutil.rmtree(macosx_folder)
+        if zip_file is not None and is_folder_empty("images_win"):
+            with zipfile.ZipFile(zip_file, "r") as zip_ref:
+                zip_ref.extractall("images_win")
+    
+    else:
+        zip_file = st.file_uploader("사진 데이터 파일 .zip 형식으로 업로드(Mac용)", type=["zip"])
+        
+        if zip_file is not None and is_folder_empty("images_mac"):
+            with zipfile.ZipFile(zip_file, "r") as zip_ref:
+                zip_ref.extractall("images_mac")
+            
+            macosx_folder = "images_mac/__MACOSX"
+            if os.path.exists(macosx_folder) and os.path.isdir(macosx_folder):
+                shutil.rmtree(macosx_folder)
 
 
-        image_folder = "images_mac"
-        folder_list = [folder for folder in os.listdir(image_folder) if os.path.isdir(os.path.join(image_folder, folder))]
+            image_folder = "images_mac"
+            folder_list = [folder for folder in os.listdir(image_folder) if os.path.isdir(os.path.join(image_folder, folder))]
 
-        for folder in folder_list:
-            folder_path = os.path.join(image_folder, folder)
-            files = os.listdir(folder_path)
-            for file in files:
-                file_path = os.path.join(folder_path, file)
-                shutil.move(file_path, image_folder)
-            os.rmdir(folder_path)
+            for folder in folder_list:
+                folder_path = os.path.join(image_folder, folder)
+                files = os.listdir(folder_path)
+                for file in files:
+                    file_path = os.path.join(folder_path, file)
+                    shutil.move(file_path, image_folder)
+                os.rmdir(folder_path)
 
 
     label_input2()
